Implementando validações com Bean Validation

+++ src/main/java/com/devsuperior/commerce/service/ProductService.java
@@ -3,11 +3,17 @@ package com.devsuperior.commerce.service;
 import com.devsuperior.commerce.dto.ProductDTO;
 import com.devsuperior.commerce.entities.Product;
 import com.devsuperior.commerce.repositories.ProductRepository;
+import com.devsuperior.commerce.service.exceptionals.DataBaseException;
 import com.devsuperior.commerce.service.exceptionals.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -37,10 +43,14 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw  new ResourceNotFoundException("Resource not found!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -49,8 +59,15 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
         entity.setPrice(dto.getPrice());
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not fount!");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException  e ){
+            throw new DataBaseException("Referencial integrity failure");
+        }
     }
 }
